+++ src/Domains/Connectors/PromptMine/Workflows/Activities/PromptVideoFilterActivity.php
@@ -10,11 +10,11 @@ use finfo;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Http\UploadedFile;
 use Illuminate\Support\Facades\Http;
-use Illuminate\Support\Facades\Log;
 use Kanvas\Apps\Models\Apps;
 use Kanvas\Companies\Models\Companies;
 use Kanvas\Companies\Models\CompaniesBranches;
 use Kanvas\Connectors\PromptMine\Actions\CreateNuggetMessageAction;
+use Kanvas\Connectors\PromptMine\Actions\ProcessVideoRequestAction;
 use Kanvas\Connectors\PromptMine\Notifications\VideoProcessingPushNotification;
 use Kanvas\Enums\AppSettingsEnums;
 use Kanvas\Exceptions\InternalServerErrorException;
@@ -46,94 +46,32 @@ class PromptVideoFilterActivity extends KanvasActivity implements WorkflowActivi
         sleep($app->get('PROMPT_VIDEO_WAIT_TIME') ?? 5);
         $this->app = $app;
 
-        // Extract video model dynamically from message - use the full value as is
-        $videoModel = $entity->message['ai_model']['value'] ?? 'fal-ai/veo3';
-        $videoType = $entity->message['type'] ?? 'video-format';
-
-        // Determine if it's text-to-video or image-to-video based on hasFiles flag
-        $isImageToVideo = isset($entity->message['hasFiles']) && $entity->message['hasFiles'] === true;
-
-        // Construct the API URL based on video type
-        $baseApiUrl = $entity->app->get('PROMPT_VIDEO_API_URL');
-        $videoKey = $isImageToVideo ? 'fal-ai/image-to-video' : 'fal-ai/text-to-video';
-        $this->apiUrl = $baseApiUrl . '/api/v2/video/' . $videoKey;
-
         $company = $this->getCompany($app, $entity);
 
         return $this->executeIntegration(
             entity: $entity,
             app: $app,
             integration: IntegrationsEnum::PROMPT_MINE,
-            integrationOperation: function ($entity, $app, $integrationCompany, $additionalParams) use ($params, $videoModel, $videoType, $isImageToVideo) {
+            integrationOperation: function ($entity, $app, $integrationCompany, $additionalParams) use ($params) {
                 $entity->setPrivate();
 
-                if (empty($this->apiUrl)) {
-                    return [
-                        'result' => false,
-                        'message' => 'Video API URL not configured',
-                    ];
-                }
-
                 try {
-                    // Check if we already have a request_id (in case of retry)
-                    $existingRequestId = $entity->message['video_request_id'] ?? null;
+                    // Use the ProcessVideoRequestAction for the core logic
+                    $processVideoAction = new ProcessVideoRequestAction($entity, $app, $params);
+                    $result = $processVideoAction->execute();
 
-                    if ($existingRequestId) {
-                        // If we already have a request ID, just return success
-                        // The delayed job will handle the processing
-                        return [
-                            'result' => true,
-                            'model' => $videoModel,
-                            'request_id' => $existingRequestId,
-                            'message' => 'Video processing request already submitted. Processing continues asynchronously.',
-                            'message_id' => $entity->getId(),
-                            'status' => 'IN_QUEUE',
-                        ];
+                    if ($result['result'] && isset($result['request_id'])) {
+                        // Schedule delayed processing
+                        $this->scheduleVideoProcessingCheck(
+                            $entity,
+                            $app,
+                            $result['request_id'],
+                            $result['model'],
+                            $params
+                        );
                     }
 
-                    if ($isImageToVideo) {
-                        // Process image-to-video
-                        $messageFiles = $entity->getFiles();
-                        if ($messageFiles->isEmpty()) {
-                            return [
-                                'result' => false,
-                                'message' => 'Message does not have any files for image-to-video processing',
-                            ];
-                        }
-
-                        $imageUrl = $messageFiles->first()->url;
-                        $requestId = $this->submitImageToVideo($imageUrl, $videoModel, $entity, $params);
-                    } else {
-                        // Process text-to-video
-                        $requestId = $this->submitTextToVideo($videoModel, $entity, $params);
-                    }
-
-                    if ($requestId === null) {
-                        return [
-                            'result' => false,
-                            'model' => $videoModel,
-                            'message' => 'Failed to submit video processing request',
-                        ];
-                    }
-
-                    // Store the request ID for tracking
-                    $messageCopy = $entity->message;
-                    $messageCopy['video_request_id'] = $requestId;
-                    $messageCopy['video_processing_status'] = 'IN_QUEUE';
-                    $entity->message = $messageCopy;
-                    $entity->save();
-
-                    // Schedule delayed processing
-                    $this->scheduleVideoProcessingCheck($entity, $app, $requestId, $videoModel, $params);
-
-                    return [
-                        'result' => true,
-                        'model' => $videoModel,
-                        'request_id' => $requestId,
-                        'message' => 'Video processing request submitted successfully. Processing will continue asynchronously.',
-                        'message_id' => $entity->getId(),
-                        'status' => 'IN_QUEUE',
-                    ];
+                    return $result;
                 } catch (Exception $e) {
                     report($e);
 
@@ -251,105 +189,6 @@ class PromptVideoFilterActivity extends KanvasActivity implements WorkflowActivi
     }
 
     /**
-     * Submit text-to-video request and return request ID
-     */
-    protected function submitTextToVideo(string $videoModel, Model $entity, array $params): ?string
-    {
-        // Get default values from app settings
-        $defaultValues = $this->getDefaultVideoValues($entity->app, 'text-to-video');
-
-        // Submit the video generation request
-        $submitPayload = [
-            'operation' => 'submit',
-            'model' => $videoModel,
-            'prompt' => $entity->message['prompt'] ?? '',
-            'resolution' => $defaultValues['resolution'] ?? '720p',
-        ];
-
-        // Add optional webhook URL if configured
-        $webhookUrl = $entity->app->get('PROMPT_VIDEO_WEBHOOK_URL');
-        if ($webhookUrl) {
-            $submitPayload['webhookUrl'] = $webhookUrl;
-        }
-
-        // Add other optional parameters based on model configuration
-        if (isset($defaultValues['aspect_ratio'])) {
-            $submitPayload['aspect_ratio'] = $defaultValues['aspect_ratio'];
-        }
-        if (isset($defaultValues['generate_audio'])) {
-            $submitPayload['generate_audio'] = $defaultValues['generate_audio'];
-        }
-        if (isset($defaultValues['duration'])) {
-            $submitPayload['duration'] = $defaultValues['duration'];
-        }
-        if (isset($defaultValues['negative_prompt'])) {
-            $submitPayload['negative_prompt'] = $defaultValues['negative_prompt'];
-        }
-        if (isset($defaultValues['style'])) {
-            $submitPayload['style'] = $defaultValues['style'];
-        }
-        if (isset($defaultValues['prompt_optimizer'])) {
-            $submitPayload['prompt_optimizer'] = $defaultValues['prompt_optimizer'];
-        }
-
-        $submitResponse = $this->submitVideoRequest($submitPayload);
-
-        if (! isset($submitResponse['request_id'])) {
-            throw new Exception('Failed to submit video for processing: ' . json_encode($submitResponse));
-        }
-
-        return $submitResponse['request_id'];
-    }
-
-    /**
-     * Submit image-to-video request and return request ID
-     */
-    protected function submitImageToVideo(string $imageUrl, string $videoModel, Model $entity, array $params): ?string
-    {
-        // Get default values from app settings
-        $defaultValues = $this->getDefaultVideoValues($entity->app, 'image-to-video');
-
-        // Submit the video generation request
-        $submitPayload = [
-            'operation' => 'submit',
-            'model' => $videoModel,
-            'image_url' => $imageUrl,
-            'prompt' => $entity->message['prompt'] ?? '',
-        ];
-
-        // Add optional webhook URL if configured
-        $webhookUrl = $entity->app->get('PROMPT_VIDEO_WEBHOOK_URL');
-        if ($webhookUrl) {
-            $submitPayload['webhookUrl'] = $webhookUrl;
-        }
-
-        // Add other optional parameters based on model configuration
-        if (isset($defaultValues['prompt_optimizer'])) {
-            $submitPayload['prompt_optimizer'] = $defaultValues['prompt_optimizer'];
-        }
-        if (isset($defaultValues['aspect_ratio'])) {
-            $submitPayload['aspect_ratio'] = $defaultValues['aspect_ratio'];
-        }
-        if (isset($defaultValues['resolution'])) {
-            $submitPayload['resolution'] = $defaultValues['resolution'];
-        }
-        if (isset($defaultValues['duration'])) {
-            $submitPayload['duration'] = $defaultValues['duration'];
-        }
-        if (isset($defaultValues['negative_prompt'])) {
-            $submitPayload['negative_prompt'] = $defaultValues['negative_prompt'];
-        }
-
-        $submitResponse = $this->submitVideoRequest($submitPayload);
-
-        if (! isset($submitResponse['request_id'])) {
-            throw new Exception('Failed to submit image-to-video for processing: ' . json_encode($submitResponse));
-        }
-
-        return $submitResponse['request_id'];
-    }
-
-    /**
      * Poll for video processing result with retries
      */
     protected function pollForVideoResult(string $requestId, string $videoModel, Apps $app): array
@@ -357,12 +196,18 @@ class PromptVideoFilterActivity extends KanvasActivity implements WorkflowActivi
         $maxAttempts = 3;
         $attempt = 0;
 
+        // Reconstruct API URL for polling
+        $isImageToVideo = isset($this->entity->message['hasFiles']) && $this->entity->message['hasFiles'] === true;
+        $baseApiUrl = $app->get('PROMPT_VIDEO_API_URL');
+        $videoKey = $isImageToVideo ? 'fal-ai/image-to-video' : 'fal-ai/text-to-video';
+        $apiUrl = $baseApiUrl . '/api/v2/video/' . $videoKey;
+
         while ($attempt < $maxAttempts) {
             try {
                 // Check status
                 $statusResponse = Http::withHeaders([
                     'Content-Type' => 'application/json',
-                ])->post($this->apiUrl, [
+                ])->post($apiUrl, [
                     'operation' => 'status',
                     'requestId' => $requestId,
                     'model' => $videoModel,
@@ -375,7 +220,7 @@ class PromptVideoFilterActivity extends KanvasActivity implements WorkflowActivi
                     // Get the result
                     $resultResponse = Http::withHeaders([
                         'Content-Type' => 'application/json',
-                    ])->post($this->apiUrl, [
+                    ])->post($apiUrl, [
                         'operation' => 'result',
                         'requestId' => $requestId,
                         'model' => $videoModel,
@@ -435,70 +280,6 @@ class PromptVideoFilterActivity extends KanvasActivity implements WorkflowActivi
             report($e);
             $this->updateVideoProcessingStatus($entity, 'FAILED', $e->getMessage());
         }
-    }
-
-    /**
-     * Get default video values from app settings
-     */
-    protected function getDefaultVideoValues(Apps $app, string $type): array
-    {
-        $settings = $app->get('llm_list_video_categorization_dev');
-
-        if (! $settings || ! is_array($settings)) {
-            return [];
-        }
-
-        $videoKey = $type === 'text-to-video' ? 'fal-ai/text-to-video' : 'fal-ai/image-to-video';
-
-        // Search through all categories for the video key
-        foreach ($settings as $category) {
-            if (! isset($category['value']) || ! is_array($category['value'])) {
-                continue;
-            }
-
-            foreach ($category['value'] as $videoTypeConfig) {
-                if (isset($videoTypeConfig['key']) && $videoTypeConfig['key'] === $videoKey) {
-                    if (isset($videoTypeConfig['value']) && is_array($videoTypeConfig['value'])) {
-                        // Find the first (default) model configuration
-                        foreach ($videoTypeConfig['value'] as $modelConfig) {
-                            if (isset($modelConfig['input_config']) && isset($modelConfig['isDefault']) && $modelConfig['isDefault']) {
-                                return $this->extractDefaultsFromInputConfig($modelConfig['input_config']);
-                            }
-                        }
-
-                        // If no default found, use the first one
-                        if (! empty($videoTypeConfig['value']) && isset($videoTypeConfig['value'][0]['input_config'])) {
-                            return $this->extractDefaultsFromInputConfig($videoTypeConfig['value'][0]['input_config']);
-                        }
-                    }
-                }
-            }
-        }
-
-        return [];
-    }
-
-    /**
-     * Extract default values from input_config
-     */
-    private function extractDefaultsFromInputConfig(array $inputConfig): array
-    {
-        $defaults = [];
-
-        foreach ($inputConfig as $key => $values) {
-            // Skip comment fields
-            if (strpos($key, '__') === 0 && strpos($key, '_comment__') !== false) {
-                continue;
-            }
-
-            if (is_array($values) && ! empty($values)) {
-                $defaults[$key] = $values[0];
-            } elseif (! is_array($values)) {
-                $defaults[$key] = $values;
-            }
-        }
-
-        return $defaults;
     }
 
     /**
@@ -622,23 +403,6 @@ class PromptVideoFilterActivity extends KanvasActivity implements WorkflowActivi
             'processed_video_url' => $processedVideoUrl,
             'request_id' => $requestId,
         ];
-    }
-
-    /**
-     * Submit a video generation request
-     */
-    protected function submitVideoRequest(array $payload): array
-    {
-        $response = Http::withHeaders([
-            'Content-Type' => 'application/json',
-        ])->post($this->apiUrl, $payload);
-
-        Log::info('Video request submitted', [
-            'url' => $this->apiUrl,
-            'payload' => $payload,
-            'response' => $response->json(),
-        ]);
-        return $response->json();
     }
 
     /**
